# swagger2.yaml
swagger: '2.0'
info:
  title: Adminiculum API
  description: Legal Workflow Platform API
  version: '1.0'
host: adminiculumaustriaeast-01.azurewebsites.net
basePath: /
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /api/v1/documents/{id}:
    get:
      summary: GetDocumentById
      description: Dokumentum részletei
      operationId: GetDocumentById
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        '200':
          description: Document details
...
securityDefinitions:
  oauth2-auth:
    type: oauth2
    flow: accessCode
    tokenUrl: https://login.microsoftonline.com/18b56834-dfea-4931-bdf8-e5ebb0cb4e0f/oauth2/v2.0/token
    authorizationUrl: https://login.microsoftonline.com/18b56834-dfea-4931-bdf8-e5ebb0cb4e0f/oauth2/v2.0/authorize
    scopes:
      api://a1e8b8a0-7690-4d09-9974-e4742d3de4e9/access_as_user: Access as user
security:
  - oauth2-auth:
      - api://a1e8b8a0-7690-4d09-9974-e4742d3de4e9/access_as_user